cambio de variable a float

+++ suma y resta.cpp
@@ -2,9 +2,11 @@
 
 int main()
 {
-	int n1,n2,resultado;
+	float n1,n2,resultado;
 	
 	printf("\n\t programa que suma y resta dos numeros");
 	
 	printf("\n\t ingresa: \n\n\t -1 para suma \n\t -2 para resta");
+	
+	
 }
